+++ src/test/java/de/uros/citlab/errorrate/KWSErrorCalculatorTest.java
@@ -6,24 +6,14 @@
 package de.uros.citlab.errorrate;
 
 import de.uros.citlab.errorrate.kws.measures.IRankingMeasure;
+import org.apache.commons.io.FileUtils;
+import org.junit.*;
+
 import java.io.File;
 import java.io.IOException;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import org.apache.commons.io.FileUtils;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import java.util.*;
 
 /**
- *
  * @author gundram
  */
 public class KWSErrorCalculatorTest {
@@ -65,25 +55,25 @@ public class KWSErrorCalculatorTest {
         File gtList = new File("src/test/resources/gt.json");
         File resFile = new File("src/test/resources/kws_htr/out_20.json");
         KwsError calculator = new KwsError();
-        Map<IRankingMeasure.Measure, Double> exp = new LinkedHashMap<>();
+        TreeMap<IRankingMeasure.Measure, Double> exp = new TreeMap<>();
         exp.put(IRankingMeasure.Measure.R_PRECISION, 0.825503355704698);
-        exp.put(IRankingMeasure.Measure.G_NCDG, -0.014978690675125528);
-        exp.put(IRankingMeasure.Measure.MAP, 0.9011214914516754);
-        exp.put(IRankingMeasure.Measure.GAP, 0.9006038330890166);
-        exp.put(IRankingMeasure.Measure.RECALL, 0.9798657718120806);
+        exp.put(IRankingMeasure.Measure.G_NCDG, -0.018474595791665722);
+        exp.put(IRankingMeasure.Measure.MAP, 0.9002287096132139);
+        exp.put(IRankingMeasure.Measure.GAP, 0.8987380224491166);
+        exp.put(IRankingMeasure.Measure.RECALL, 0.9765100671140939);
         exp.put(IRankingMeasure.Measure.PRECISION_AT_10, 0.5054545454545453);
-        exp.put(IRankingMeasure.Measure.PRECISION, 0.4165477888730385);
-        exp.put(IRankingMeasure.Measure.M_NCDG, 0.38384328267815787);
-        exp.put(IRankingMeasure.Measure.WMAP, 0.915926584620434);
+        exp.put(IRankingMeasure.Measure.PRECISION, 0.41512125534950073);
+        exp.put(IRankingMeasure.Measure.M_NCDG, 0.3827312294124316);
+        exp.put(IRankingMeasure.Measure.WMAP,0.9141140577336244);
         Map<IRankingMeasure.Measure, Double> run = calculator.run(new String[]{resFile.getPath(), gtList.getPath()});
         for (IRankingMeasure.Measure measure : run.keySet()) {
 //            System.out.println(measure + " = " + run.get(measure));
-            Assert.assertEquals("measure changed", exp.get(measure), run.get(measure), 0.000001);
+            Assert.assertEquals("measure changed (" + measure + ")", exp.get(measure), run.get(measure), 0.000001);
         }
         // TODO review the generated test code and remove the default call to fail.
     }
 
-    public static void main(String[] args) throws IOException {
+    public static void min(String[] args) throws IOException {
         KWSErrorCalculatorTest.setUpClass();
         KWSErrorCalculatorTest t = new KWSErrorCalculatorTest();
         t.testRun();