+++ src/test/java/de/uros/citlab/errorrate/kws/KWSEvaluationMeasureTest.java
@@ -7,42 +7,27 @@ package de.uros.citlab.errorrate.kws;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.panayotis.gnuplot.JavaPlot;
 import de.uros.citlab.errorrate.kws.measures.IRankingMeasure;
-import de.uros.citlab.errorrate.aligner.BaseLineAligner;
 import de.uros.citlab.errorrate.kws.measures.IRankingStatistic;
 import de.uros.citlab.errorrate.types.KWS;
 import de.uros.citlab.errorrate.types.KWS.Line;
 import de.uros.citlab.errorrate.types.KWS.Page;
 import de.uros.citlab.errorrate.types.KWS.Result;
 import de.uros.citlab.errorrate.types.KWS.Word;
-import de.uros.citlab.errorrate.util.PlotUtil;
-import java.awt.Polygon;
+import org.apache.commons.io.FileUtils;
+import org.junit.*;
+
+import java.awt.*;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Random;
-import java.util.Set;
-import java.util.function.Consumer;
-import org.apache.commons.io.FileUtils;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import java.util.*;
+
 import static org.junit.Assert.*;
 
 /**
- *
  * @author gundram
  */
 public class KWSEvaluationMeasureTest {
@@ -83,18 +68,17 @@ public class KWSEvaluationMeasureTest {
 //    @Test
     public void testGetGlobalMearsure() {
         System.out.println("getGlobalMearsure");
-        KWSEvaluationMeasure measure = new KWSEvaluationMeasure(new BaseLineAligner());
 
-        test(measure, 1.0, 0.0);
-        test(measure, 0.8, 0.0);
-        test(measure, 0.5, 0.0);
-        test(measure, 0.0, 0.0);
+        test(1.0, 0.0);
+        test(0.8, 0.0);
+        test(0.5, 0.0);
+        test(0.0, 0.0);
 
-        test(measure, 1.0, 0.2);
-        test(measure, 1.0, 1.0);
+        test(1.0, 0.2);
+        test(1.0, 1.0);
     }
 
-    private void test(KWSEvaluationMeasure measure, double corrRatio, double fnRatio) {
+    private void test(double corrRatio, double fnRatio) {
         HashMap<String, KWS.Word> words = new HashMap<>();
         List<KWS.Page> pages = new LinkedList<>();
         int numOfPages = 1;
@@ -125,7 +109,7 @@ public class KWSEvaluationMeasureTest {
                 totalcorr += numOfcorr;
                 totalFn += numOfFn;
                 Polygon p = new Polygon(new int[]{lineId * 50, lineId * 50}, new int[]{0, 100}, 2);
-                Line line = new Line(p);
+                Line line = new Line("line_" + lineId, p, null);
 
                 int cnt = -1;
                 Collections.shuffle(keyAndWord, rnd);
@@ -156,23 +140,22 @@ public class KWSEvaluationMeasureTest {
         Result res = new Result(new HashSet<>(words.values()));
         KWS.GroundTruth gt = new KWS.GroundTruth(pages);
 
-        measure.setGroundtruth(gt);
-        measure.setResults(res);
         LinkedList<IRankingMeasure.Measure> ms = new LinkedList<>();
         ms.add(IRankingMeasure.Measure.GAP);
 
-        double globalMearsure = measure.getMeasure(ms).get(IRankingMeasure.Measure.GAP);
+        double globalMearsure = KWSEvaluationMeasure.getMeasure(res, gt, KeyWordMatchers.nearBaselines(), ms).get(IRankingMeasure.Measure.GAP);
         assertEquals(corrRatio == 0.0 ? 1.0 : (double) totalcorr / (totalcorr + totalFn), globalMearsure, 1e-5);
         System.out.println("measure: " + globalMearsure);
     }
 
     private void addMatch(Word word, double conf, Polygon p, String pageId) {
-        word.add(new KWS.Entry(conf, "", p, pageId));
+        word.add(new KWS.Entry(conf, "", pageId, p, null));
     }
 
     private void addGtWord(Line line, String keyword, Polygon p, String pageId) {
-        line.addKeyword(keyword, p);
+        line.addKeyword(keyword, p, null);
     }
+
     private static final File folderGT = new File("src/test/resources/gt");
     private static File[] listGT;
 
@@ -226,15 +209,15 @@ public class KWSEvaluationMeasureTest {
         List<String> readLines = FileUtils.readLines(new File("src/test/resources/kw.txt"));
 //        List<String> readLines = Arrays.asList("seyn");
         KeywordExtractor kwe = new KeywordExtractor();
-        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(getStringList(listGT), null, readLines);
-        KWSEvaluationMeasure kem = new KWSEvaluationMeasure(new BaseLineAligner());
-        kem.setGroundtruth(keywordGroundTruth);
+        KeywordExtractor.PageIterator pi = new KeywordExtractor.FileListPageIterator(getStringList(listGT), null);
+        KeywordExtractor.KeyWordProvider kp = new KeywordExtractor.FixedKeyWordProvider(readLines);
+
+        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(pi, kp);
         LinkedList<KWS.MatchList> mls = new LinkedList<>();
         for (int i : new int[]{50, 20, 10, 5}) {
             Result res = getResult(new File(String.format("src/test/resources/kws_htr/out_%02d.json", i)));
             res = filter(res, readLines);
-            kem.setResults(res);
-            mls.add(merge(kem.getMatchList()));
+            mls.add(merge(KWSEvaluationMeasure.getMatchList(res, keywordGroundTruth, KeyWordMatchers.nearBaselines())));
         }
         for (int i = 1; i < mls.size(); i++) {
             if (mls.get(i - 1).matches.size() > mls.get(i).matches.size()) {
@@ -251,8 +234,8 @@ public class KWSEvaluationMeasureTest {
             for (int idxMatch = 0; idxMatch < listSmall.matches.size(); idxMatch++) {
                 KWS.Match matchSmallList = listSmall.matches.get(idxMatch);
                 KWS.Match matchLargeList = listLarge.matches.get(idxMatch);
-                if (matchSmallList.conf != matchLargeList.conf && matchSmallList.type != KWS.Type.FALSE_NEGATIVE) {
-                    assertTrue("confidences differ on idxMatch " + idxMatch + " and between indexes " + (idxList2 - 1) + " and " + idxList2 + ".", matchSmallList.conf == matchLargeList.conf
+                if (matchSmallList.getHypConfidence() != matchLargeList.getHypConfidence() && matchSmallList.type != KWS.Type.FALSE_NEGATIVE) {
+                    assertTrue("confidences differ on idxMatch " + idxMatch + " and between indexes " + (idxList2 - 1) + " and " + idxList2 + ".", matchSmallList.matchConf == matchLargeList.matchConf
                     );
                 }
             }
@@ -266,19 +249,19 @@ public class KWSEvaluationMeasureTest {
         List<String> readLines = FileUtils.readLines(new File("src/test/resources/kw.txt"));
 //        List<String> readLines = Arrays.asList("seyn");
         KeywordExtractor kwe = new KeywordExtractor();
-        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(getStringList(listGT), null, readLines);
-        KWSEvaluationMeasure kem = new KWSEvaluationMeasure(new BaseLineAligner());
-        kem.setGroundtruth(keywordGroundTruth);
+        KeywordExtractor.PageIterator pi = new KeywordExtractor.FileListPageIterator(getStringList(listGT), null);
+        KeywordExtractor.KeyWordProvider kp = new KeywordExtractor.FixedKeyWordProvider(readLines);
+
+        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(pi, kp);
         IRankingMeasure.Measure[] ms = new IRankingMeasure.Measure[]{
-            IRankingMeasure.Measure.GAP, IRankingMeasure.Measure.MAP,
-            IRankingMeasure.Measure.R_PRECISION, IRankingMeasure.Measure.PRECISION,
-            IRankingMeasure.Measure.RECALL, IRankingMeasure.Measure.PRECISION_AT_10, IRankingMeasure.Measure.WMAP};
+                IRankingMeasure.Measure.GAP, IRankingMeasure.Measure.MAP,
+                IRankingMeasure.Measure.R_PRECISION, IRankingMeasure.Measure.PRECISION,
+                IRankingMeasure.Measure.RECALL, IRankingMeasure.Measure.PRECISION_AT_10, IRankingMeasure.Measure.WMAP};
         for (IRankingMeasure.Measure m : ms) {
             for (int i : new int[]{5, 10, 20, 50}) {
                 Result res = getResult(new File(String.format("src/test/resources/kws_htr/out_%02d.json", i)));
                 res = filter(res, readLines);
-                kem.setResults(res);
-                Map<IRankingMeasure.Measure, Double> measure = kem.getMeasure(m);
+                Map<IRankingMeasure.Measure, Double> measure = KWSEvaluationMeasure.getMeasure(res, keywordGroundTruth, KeyWordMatchers.nearBaselines(),m);
                 System.out.println("#### i = " + i + " ####");
                 for (IRankingMeasure.Measure measure1 : measure.keySet()) {
                     System.out.println(measure1.toString() + " = " + measure.get(measure1));
@@ -301,9 +284,10 @@ public class KWSEvaluationMeasureTest {
         List<String> readLines = FileUtils.readLines(new File("src/test/resources/kw.txt"));
 //        List<String> readLines = Arrays.asList("sein");
         KeywordExtractor kwe = new KeywordExtractor();
-        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(getStringList(listGT), null, readLines);
-        KWSEvaluationMeasure kem = new KWSEvaluationMeasure(new BaseLineAligner());
-        kem.setGroundtruth(keywordGroundTruth);
+        KeywordExtractor.PageIterator pi = new KeywordExtractor.FileListPageIterator(getStringList(listGT));
+        KeywordExtractor.KeyWordProvider kp = new KeywordExtractor.FixedKeyWordProvider(readLines);
+
+        KWS.GroundTruth keywordGroundTruth = kwe.getKeywordGroundTruth(pi, kp);
         List<double[]> data = new LinkedList<>();
         String[] names = new String[8];
         int idx = 0;
@@ -313,9 +297,8 @@ public class KWSEvaluationMeasureTest {
             File filename = new File(String.format("src/test/resources/kws_htr/out_%02d.json", i));
             Result res = getResult(filename);
             res = filter(res, readLines);
-            kem.setResults(res);
             List<IRankingStatistic.Statistic> asList = Arrays.asList(IRankingStatistic.Statistic.W_PR_CURVE, IRankingStatistic.Statistic.PR_CURVE);
-            Map<IRankingStatistic.Statistic, double[]> stats = kem.getStats(asList);
+            Map<IRankingStatistic.Statistic, double[]> stats = KWSEvaluationMeasure.getStats(res, keywordGroundTruth, KeyWordMatchers.nearBaselines(), asList);
             System.out.println("#### i = " + i + " ####");
             for (IRankingStatistic.Statistic measure1 : stats.keySet()) {
                 System.out.println(measure1.toString() + "========================= " + Arrays.toString(stats.get(measure1)));